requirements.txt
flask==2.3.3
requests==2.31.0
gunicorn==21.2.0
pyyaml==6.0
python-dotenv==1.0.0
openai==1.3.0

# Add IMAPClient for better IMAP support
imapclient==2.3.1
